pyproject.toml
[project]
name = "biodiversity-chatbot"
version = "0.1.0"
description = "Biodiversity Chatbot"
readme = "README.md"
requires-python = ">= 3.12"
authors = [
    { name = "Csaba Boros", email = "csaba.boros@outlook.com" }
]
dependencies = [
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "google-cloud-bigquery>=3.27.0",
    "setuptools>=75.6.0",
    "pygbif>=0.6.5",
    "langchain==0.3.9",
    "langchain-community==0.3.9",
    "langchain_google_community>=2.0.3",
    "EcoNameTranslator>=2.1",
    "streamlit>=1.42.0",
    "scipy>=1.11.4",
    "shapely>=1.8.3",
    "scikit-learn>=1.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.22.10",
    "plotly>=5.22.0",
    "google-cloud-aiplatform>=1.83.0",
    "db-dtypes>=1.0.0",
    "google-cloud-bigquery-storage>=2.0.0",
    "earthengine-api>=0.1.375",
    "tqdm>=4.66.1",
    "python-dotenv>=1.0.0",
    "pydeck-earthengine-layers>=0.2.0",
    "streamlit-folium>=0.12.0",
    "pycountry>=23.1.15",
    "google-cloud-secret-manager>=2.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pytest-xdist>=3.6.1",
    "pytest-mock>=3.14.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "app/**/*.py",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = "app/tests"
addopts = "-v -s"
